refactor(recognition): extract recognition tauri layer into ts-recognition-tauri (#355)
Вынос Tauri-командного слоя распознавания из монолита `src-tauri/src/recognition/` в крейт
`ts-recognition-tauri`: команды YOLO/FaceNet/RetinaFace/MediaPipe/privacy/clustering/
person/advanced_tracking + vision_service (оркестратор). Инференс-движок уже в `ts-recognition`
(#96), реэкспортируется отсюда.

Разблокировано выносом state (#360):
- crate::state::project_state -> ts_state_tauri::project_state; crate::state -> ts_state_tauri
- crate::analysis::types -> ts_analysis::analysis::types; crate::recognition:: (intra) -> crate::
- privacy_commands: use crate::commands::yolo_commands::YoloProcessorState (точный путь).

Монолит: `recognition/` -> файл-шим `src/recognition.rs = pub use ts_recognition_tauri::*`
(montage_planner, media, app_builder резолвятся без правок).
Зависимости: ts-recognition, ts-state-tauri, ts-analysis, tauri, image, base64, dirs, ...

Проверки: cargo test -p ts-recognition-tauri --test-threads=2 -> 13 passed;
cargo check --lib (src-tauri) -> 0 ошибок, 3 baseline-варнинга.

Closes #355. Часть #354.
