+++ crates/ts-recognition-tauri/src/README.ru.md
